+++ nengo_edge/device_modules/tests/test_coral_device.py
@@ -11,7 +11,6 @@ import pytest
 
 from nengo_edge import network_runner
 from nengo_edge.device_modules import coral_device, np_mfcc
-from nengo_edge.tests.test_network_runner import MockRunner
 
 
 class MockDeviceRunner(coral_device.CoralDeviceRunner):
@@ -48,7 +47,6 @@ def test_cli(
     batch_size = 3
     inputs = rng.uniform(-0.5, 0.5, size=(batch_size, 16000)).astype("float32")
 
-    monkeypatch.setattr(network_runner, "CoralRunner", MockRunner)
     monkeypatch.setattr(coral_device, "CoralDeviceRunner", MockDeviceRunner)
     monkeypatch.setattr(
         coral_device, "_build_feature_extractor", mock_feature_extractor
@@ -57,17 +55,14 @@ def test_cli(
     binary_path = param_dir / "model_edgetpu.tflite"
     binary_path.touch()
 
-    runner = MockRunner(directory=param_dir, username="user", hostname="hostname")
+    runner = network_runner.CoralRunner(
+        directory=param_dir, username="user", hostname="hostname", local=True
+    )
 
     # get output from cli
-    assert isinstance(runner.remote_dir, Path)
-    runner.prepare_device_runner()
-    runner.send_inputs(inputs)
-
+    np.savez_compressed(runner.remote_dir / "inputs.npz", inputs=inputs)
     monkeypatch.setattr(
-        sys,
-        "argv",
-        f"coral_runner.py --directory {runner.remote_dir}".split(),
+        sys, "argv", f"coral_runner.py --directory {runner.remote_dir}".split()
     )
     coral_device.cli()