+++ nengo_edge/network_runner.py
@@ -5,8 +5,9 @@ Nengo-edge supports running on the Coral dev board via this runner.
 """
 
 import subprocess
+import tempfile
 from pathlib import Path
-from typing import List, Union
+from typing import List, Optional, Union
 
 import numpy as np
 
@@ -16,93 +17,102 @@ from nengo_edge.device_modules import coral_device, np_mfcc
 
 class NetworkRunner:
     """
-    Run an exported model from nengo edge that runs on remote network devices.
+    Run an exported component on a remote network device.
 
     Parameters
     ----------
     directory : Union[str, Path]
-        Path to the directory containing the nengo edge export artifacts.
-    device_runner : Union[str, Path]
-        Path to the file that will run on the device.
+        Path to the directory containing the NengoEdge export artifacts.
     username : str
         Username on the remote device
     hostname : str
         Hostname of the remote device
+    local : bool
+        If True, run locally rather than over the network (for debugging purposes).
     """
 
     def __init__(
         self,
         directory: Union[str, Path],
-        device_runner: Union[str, Path],
         username: str,
         hostname: str,
+        local: bool = False,
     ):
         self.directory = Path(directory)
-        self.device_runner = Path(device_runner)
         self.model_params, self.preprocessing, self.postprocessing = config.load_params(
             self.directory
         )
-        self.return_sequences = self.model_params["return_sequences"]
-
         self.username = username
         self.hostname = hostname
         self.address = f"{self.username}@{self.hostname}"
-        self.check_connection()
+        self.local = local
 
-        self.remote_dir = Path("/tmp/nengo-edge-runner")
-
-        self.reset()
-
-    def check_connection(self) -> None:
-        """
-        Run ssh to confirm connection to remote device.
-
-        Throws runtime error if connection fails.
-        """
         try:
-            subprocess.run(f"ssh {self.address} echo ok".split(), check=True)
+            self._ssh("echo ok")
         except Exception as e:
             raise RuntimeError(
                 f"Cannot connect to address {self.address}: {e!r}"
             ) from e
 
-    def _scp(
-        self, files_to_copy: List[Path]
-    ) -> None:  # pragma: no cover (needs device)
+        if local:
+            self._tmp_dir = (
+                tempfile.TemporaryDirectory(  # pylint: disable=consider-using-with
+                    dir=directory
+                )
+            )
+            self.remote_dir = Path(self._tmp_dir.name)
+        else:  # pragma: no cover (needs device)
+            self.remote_dir = Path(f"/tmp/nengo-edge-{type(self).__name__.lower()}")
+            self._ssh(f"rm -rf {self.remote_dir}")
+            self._ssh(f"mkdir -p {self.remote_dir}")
+
+    def _scp(self, files_to_copy: List[Path]) -> None:
         """One liner to send specified files to remote device location."""
         cmd = (
-            f"scp {' '.join(str(p) for p in files_to_copy)} "
-            f"{self.address}:{self.remote_dir}"
+            (["cp", "-r"] if self.local else ["scp"])
+            + [str(p) for p in files_to_copy]
+            + [
+                (
+                    str(self.remote_dir)
+                    if self.local
+                    else f"{self.address}:{self.remote_dir}"
+                )
+            ]
         )
 
-        subprocess.run(cmd.split(), check=True)
+        subprocess.run(cmd, check=True)
 
-    def reset(self) -> None:
-        """Reset the state of the network runner."""
-        self.prepared = False
+    def _ssh(self, cmd: str, std_in: Optional[bytes] = None) -> bytes:
+        """Run a command over ssh."""
+        return subprocess.run(
+            cmd if self.local else ["ssh", self.address, cmd],
+            input=std_in,
+            check=True,
+            stdout=subprocess.PIPE,
+            shell=self.local,
+        ).stdout
 
-    def send_inputs(self, inputs: np.ndarray) -> None:
-        """Saves inputs to file and sends to device."""
 
-        # save inputs to file
-        filepath = self.directory / "inputs.npz"
-        np.savez_compressed(filepath, inputs=inputs)
+class CoralRunner(NetworkRunner):
+    """
+    Run a model exported from NengoEdge on the Coral board.
 
-        # copy to device
-        self._scp([filepath])
+    See `NetworkRunner` for parameter descriptions.
+    """
 
-    def prepare_device_runner(self) -> None:  # pragma: no cover (needs device)
-        """Send required runtime parameters/modules before any inputs."""
-        assert (
-            len(self.remote_dir.parts) > 2 and self.remote_dir.parts[1] == "tmp"
-        ), f"remote_dir ({self.remote_dir}) not in /tmp"
-
-        subprocess.run(
-            f"ssh {self.address} rm -rf {self.remote_dir}".split(), check=True
+    def __init__(
+        self,
+        directory: Union[str, Path],
+        username: str,
+        hostname: str,
+        local: bool = False,
+    ):
+        super().__init__(
+            directory=directory, username=username, hostname=hostname, local=local
         )
-        subprocess.run(
-            f"ssh {self.address} mkdir -p {self.remote_dir}".split(), check=True
-        )
+
+        self.device_runner = Path(coral_device.__file__)
+        self.return_sequences = self.model_params["return_sequences"]
 
         # copy files to remote
         self._scp(
@@ -113,32 +123,11 @@ class NetworkRunner:
                 Path(np_mfcc.__file__),
             ]
         )
-        self.prepared = True
 
-    def _run_model(
+    def run(
         self,
         inputs: np.ndarray,
     ) -> np.ndarray:  # pragma: no cover (needs device)
-        """Run the main model logic on the given inputs."""
-        self.send_inputs(inputs)
-        subprocess.run(
-            f"ssh {self.address} python3 {self.remote_dir / self.device_runner.name} "
-            f"--directory {self.remote_dir} "
-            f"{'--return-sequences' if self.return_sequences else ''}".split(),
-            check=True,
-        )
-
-        # copy outputs back to host
-        subprocess.run(
-            f"scp {self.address}:{self.remote_dir / 'outputs.npy'} "
-            f"{self.directory}".split(),
-            check=True,
-        )
-
-        outputs = np.load(self.directory / "outputs.npy")
-        return outputs
-
-    def run(self, inputs: np.ndarray) -> np.ndarray:  # pragma: no cover (needs device)
         """
         Run model inference on a batch of inputs.
 
@@ -154,29 +143,27 @@ class NetworkRunner:
             ``self.model_params['return_sequences']=False``
             else ``(batch_size, output_steps, output_d)``).
         """
-        if not self.prepared:
-            self.prepare_device_runner()
 
-        outputs = self._run_model(inputs)
-        return outputs
+        # Save inputs to file
+        filepath = self.directory / "inputs.npz"
+        np.savez_compressed(filepath, inputs=inputs)
 
+        # Copy to device
+        self._scp([filepath])
 
-class CoralRunner(NetworkRunner):
-    """
-    Run a model exported from NengoEdge on the Coral board.
-
-    See `NetworkRunner` for parameter descriptions.
-    """
-
-    def __init__(
-        self,
-        directory: Union[str, Path],
-        username: str,
-        hostname: str,
-    ):
-        super().__init__(
-            directory=directory,
-            username=username,
-            hostname=hostname,
-            device_runner=coral_device.__file__,
+        # Run model on device
+        self._ssh(
+            f"python3 {self.remote_dir / self.device_runner.name} "
+            f"--directory {self.remote_dir} "
+            f"{'--return-sequences' if self.return_sequences else ''}"
         )
+
+        # Copy outputs back to host
+        subprocess.run(
+            f"scp {self.address}:{self.remote_dir / 'outputs.npy'} "
+            f"{self.directory}".split(),
+            check=True,
+        )
+
+        outputs = np.load(self.directory / "outputs.npy")
+        return outputs